+++ code/src/App.js
@@ -1,24 +1,10 @@
-import React from 'react';
-import { Provider } from 'react-redux';
-import { combineReducers, configureStore } from '@reduxjs/toolkit';
-import tasks from 'reducers/tasks';
-import TaskList from 'components/TaskList';
+import React from 'react'
+import { ToDoWrapper } from 'components/ToDoWrapper'
 
 export const App = () => {
-  const reducer = combineReducers({
-    tasks: tasks.reducer
-  });
-
-  const store = configureStore({
-    // reducer: reducer
-    reducer
-  });
-
   return (
-    <Provider store={store}>
-
-      <TaskList />
-
-    </Provider>
-  );
+    <div>
+      <ToDoWrapper />
+    </div>
+  )
 }
