+++ code/src/components/Header.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <h1>To-do</h1>
+  )
+}
+
+export default Header;
